consolidate some options tests
